--- 30412/Vrabie Diana Teodora/Assignment 3 Extra/parseData.h
@@ -1,16 +1,0 @@
-//
-// Created by diana on 07/05/2020.
-//
-
-#ifndef ASSIGNMENT_3_EXTRA_PARSEDATA_H
-#define ASSIGNMENT_3_EXTRA_PARSEDATA_H
-
-#include <stdio.h>
-#include <stdlib.h>
-#include "queue.h"
-
-#define MAX_CHAR_LENGTH 6
-
-void readQueue(FILE *inputFile, Queue *queue);
-
-#endif //ASSIGNMENT_3_EXTRA_PARSEDATA_H
