+++ 30412/Vrabie Diana Teodora/Assignment 3 Extra/queue.h
@@ -9,6 +9,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <stdbool.h>
+#define MAX_CHAR_LENGTH 6
 
 typedef struct _queue {
     QueueNode *first, *last;
@@ -21,5 +22,7 @@ void enqueue(QueueNode *node, Queue *queue);
 QueueNode *dequeue (Queue *queue);
 
 bool isEmpty (Queue queue);
+
+void readQueue(FILE *inputFile, Queue *queue);
 
 #endif //ASSIGNMENT_3_EXTRA_QUEUE_H
